+++ modules/news/src/admin/messages/de/newsadmin.php
@@ -6,12 +6,10 @@ return [
 
     'news' => 'News',
     'news_administrate' => 'Verwalten',
-
-    //article
     'article' => 'News Eintrag',
     'article_cat_id' => 'Kategorie',
     'article_title' => 'Titel',
-    'article_text' => 'Beschreibung',
+    'article_text' => 'Inhalt',
     'article_image_id' => 'Bild',
     'article_timestamp_create' => 'Erstellungsdatum',
     'article_timestamp_display_from' => 'Artikel anzeigen ab',
@@ -20,15 +18,15 @@ return [
     'article_image_list' => 'Bilderliste',
     'article_file_list' => 'Dateiliste',
     'article_tag' => 'Tag-Name',
-
-    //cat
     'cat' => 'Kategorien',
     'cat_title' => 'Kategorie',
     'cat_delete_error' => 'Diese Kategorie wird noch von einem oder mehreren Terminen benutzt und kann nicht gelöscht werden.',
     'cat_title_create_error' => 'Bitte geben Sie einen Kategorienamen ein.',
-
-    //tag
     'tag' => 'Tags',
     'tag_title' => 'Tag-Name',
     'tag_title_create_error' => 'Bitte geben Sie einen Tag-Namen ein.',
+    
+// 1.0.0
+    
+    'teaser_text' => 'Teaser Text',
 ];