+++ modules/news/src/admin/messages/en/newsadmin.php
@@ -6,12 +6,10 @@ return [
 
     'news' => 'News',
     'news_administrate' => 'Administrate',
-
-    //article
     'article' => 'News entry',
     'article_cat_id' => 'Category',
     'article_title' => 'Title',
-    'article_text' => 'Description',
+    'article_text' => 'Content',
     'article_image_id' => 'Image',
     'article_timestamp_create' => 'Date',
     'article_timestamp_display_from' => 'Display from ',
@@ -20,15 +18,15 @@ return [
     'article_image_list' => 'Image list',
     'article_file_list' => 'File list',
     'article_tag' => 'Tag',
-
-    //cat
     'cat' => 'Categories',
     'cat_title' => 'Category',
     'cat_delete_error' => 'This category is still being in use by one or multiple news and can not be deleted.',
     'cat_title_create_error' => 'Please choose a category title.',
-
-    //tag
     'tag' => 'Tags',
     'tag_title' => 'Tag',
     'tag_title_create_error' => 'Please choose a tag title.',
+    
+// 1.0.0
+    
+    'teaser_text' => 'Teaser Text',
 ];