+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The changelog contains informations about bug fixes, new features or bc breaking
 
 ### Added
 
+- [#1096](https://github.com/luyadev/luya/issues/1096) New teaser_text field for news module article model.
 - [#1006](https://github.com/luyadev/luya/issues/1006) Added spanish translations to all luya core modules and administration interface.
 
 ### Fixed